requirements.txt
streamlit
langchain-groq
langchain_community
python-dotenv
langchain_core
langchain_ollama

langserve
fastapi
uvicorn
sse_starlette

sentence-transformers
pypdf
bs4
chromadb
faiss-cpu

arxiv
wikipedia
langchainhub
requests
beautifulsoup4
googlesearch-python

Flask

streamlit_webrtc
SpeechRecognition
gtts
sounddevice 
scipy 

crewai
langchain_openai
tavily-python
crewai_tools
